POSUI-244 BroadPOS payment app's security keyboard is covering the POSUI apps toast error message for incorrect password entry

+++ app/src/main/java/com/paxus/pay/poslinkui/demo/entry/text/fsa/FSAFragment.java
@@ -209,7 +209,7 @@ public class FSAFragment extends BaseEntryFragment {
     }
 
     @Override
-    protected void onEntryDeclined(long errCode, String errMessage) {
+    protected void onEntryDeclined(int errCode, String errMessage) {
         super.onEntryDeclined(errCode, errMessage);
 
         //Re enter sub health care amount
